requirements.txt
Flask
Pillow
keras
numpy
requests
gunicorn
tensorflow-cpu
